requirements.txt
gradio>=4.0.0
huggingface_hub
numpy
opencv-python
pillow